Many updates

+++ controllers/client/users.controller.js
@@ -4,39 +4,34 @@ const jwt = require('jsonwebtoken')
 const fs = require('fs')
 require('dotenv').config()
 const AsyncHandler = require("express-async-handler")
+const check = require('./check')
 
 exports.login = AsyncHandler(async (req, res) => {
     try {
-        if (!req.body.email) {
-            res.status(400).json({ msg: "Email input is empty" });
-        } else
-            if (!req.body.password) {
-                res.status(400).json({ msg: "Password input is empty" });
+        if (check.variables(['email', 'password'], req.body, res)) {
+            const user = await models.Users.findOne({ where: { email: req.body.email } })
+            if (!user) {
+                res.status(404).json({ msg: "User is not define" });
+            } else if (user.ban) {
+                res.status(403).json({ msg: 'You are banned' })
+            } else if (!await bcrypt.compare(req.body.password, user.password)) {
+                return res.status(400).json({ msg: "Password is wrong" })
             } else {
-                const user = await models.Users.findOne({ where: { email: req.body.email } })
-                if (!user) {
-
-                    res.status(404).json({ msg: "User is not define" });
-                } else if (user.ban) {
-                    res.status(403).json({ msg: 'You are banned' })
-                } else if (!await bcrypt.compare(req.body.password, user.password)) {
-                    return res.status(400).json({ msg: "Password is wrong" })
-                } else {
-                    const token = await jwt.sign(
-                        { user_id: user.id, type: user.type },
-                        process.env.TOKEN_KEY,
-                        {
-                            expiresIn: "24h",
-                        }
-                    )
-                    res.status(200).json({
-                        image: user.img,
-                        nick: user.nick,
-                        type: user.type,
-                        token: token
-                    })
-                }
+                const token = await jwt.sign(
+                    { user_id: user.id, type: user.type },
+                    process.env.TOKEN_KEY,
+                    {
+                        expiresIn: "24h",
+                    }
+                )
+                res.status(200).json({
+                    image: user.img,
+                    nick: user.nick,
+                    type: user.type,
+                    token: token
+                })
             }
+        }
     } catch (error) {
         console.log(error)
     }
@@ -45,48 +40,38 @@ exports.login = AsyncHandler(async (req, res) => {
 
 exports.registry = async (req, res) => {
     try {
-        if (!req.body.email) {
-            res.status(400).json({ msg: "Email input is empty" });
-        }
-        if (!req.body.password) {
-            res.status(400).json({ msg: "Password input is empty" });
-        }
-        if (!req.body.name) {
-            res.status(400).json({ msg: "Name input is empty" });
-        }
-        if (!req.body.surname) {
-            res.status(400).json({ msg: "Surname input is empty" });
-        }
-        const oldUser = await models.Users.findOne({
-            where: {
-                email: req.body.email
+        if (check.variables(['email', 'password', 'name', 'surname'], req.body, res)) {
+            const oldUser = await models.Users.findOne({
+                where: {
+                    email: req.body.email
+                }
+            })
+            if (oldUser) {
+                res.status(400).json({ msg: "Email Already Exist" });
+            } else {
+                const encryptedUserPassword = await bcrypt.hash(req.body.password, 10);
+                const user = await models.Users.create({
+                    nick: `${req.body.name} ${req.body.surname}`,
+                    name: req.body.name,
+                    surname: req.body.surname,
+                    email: req.body.email.toLowerCase(),
+                    password: encryptedUserPassword,
+                });
+                const token = jwt.sign(
+                    { user_id: user.id, type: user.type },
+                    process.env.TOKEN_KEY,
+                    {
+                        expiresIn: "5h",
+                    }
+                );
+                res.json({
+                    image: user.img,
+                    nick: user.nick,
+                    type: user.type,
+                    token: token
+                });
             }
-        })
-        if (oldUser) {
-            return res.status(409).json({ msg: "Email Already Exist" });
         }
-        const encryptedUserPassword = await bcrypt.hash(req.body.password, 10);
-        const user = await models.Users.create({
-            nick: `${req.body.name} ${req.body.surname}`,
-            name: req.body.name,
-            surname: req.body.surname,
-            email: req.body.email.toLowerCase(), // sanitize
-            password: encryptedUserPassword,
-        });
-        const token = jwt.sign(
-            { user_id: user.id, type: user.type },
-            process.env.TOKEN_KEY,
-            {
-                expiresIn: "5h",
-            }
-        );
-        let body = {
-            image: user.img,
-            nick: user.nick,
-            type: user.type,
-            token: token
-        }
-        res.status(200).json(body);
     } catch (error) {
         console.log(error)
     }
@@ -96,45 +81,41 @@ exports.registry = async (req, res) => {
 
 exports.cabinet = async (req, res) => {
     try {
-
-        if (!req.params.id) {
-            res.status(400).json({
-                msg: "Id input is empty"
+        if (check.variables(['id'], req.params, res)) {
+            let user = await models.Users.findOne({
+                where: {
+                    id: req.params.id
+                }
             })
-        }
-        let user = await models.Users.findOne({
-            where: {
-                id: req.params.id
-            }
-        })
-        if (!user) {
-            res.status(404).json({
-                msg: 'Account not define'
-            })
-        }
-        let data
-        if (req.params.id == req.id) {
-            data = {
-                nick: user.nick,
-                name: user.name,
-                surname: user.surname,
-                img: user.img,
-                info: user.info,
-                ball: user.ball,
-                email: user.email,
-                type: process.env.Types.split(' ')[user.type]
-            }
-        } else {
-            data = {
-                nick: user.nick,
-                img: user.img,
-                info: user.info,
-                ball: user.ball,
-                email: user.email
+            if (!user) {
+                res.status(404).json({
+                    msg: 'Account not define'
+                })
+            } else {
+                let data
+                if (req.params.id == req.id) {
+                    data = {
+                        nick: user.nick,
+                        name: user.name,
+                        surname: user.surname,
+                        img: user.img,
+                        info: user.info,
+                        ball: user.ball,
+                        email: user.email,
+                        type: process.env.Types.split(' ')[user.type]
+                    }
+                } else {
+                    data = {
+                        nick: user.nick,
+                        img: user.img,
+                        info: user.info,
+                        ball: user.ball,
+                        email: user.email
+                    }
+                }
+                res.status(200).json(data)
             }
         }
-        res.status(200).json(data)
-
     } catch (error) {
         console.log(error)
     }
@@ -142,35 +123,38 @@ exports.cabinet = async (req, res) => {
 
 exports.image = async (req, res) => {
     try {
-        let user = await models.Users.findOne({
-            where: {
-                id: req.id
-            }
-        })
-        if (!user) {
-            res.status(404).json({
-                msg: 'Account not define'
-            })
-        }
-        const old = user.img
-        if (user.img !== "public/images/default_avatar.jpg") {
-            fs.unlink(old, (err) => {
-                if (err) return console.log(err)
-            });
-        }
-        await models.Users.update(
-            {
-                img: req.file.path
-            },
-            {
+        if (check.variables(['id'], req, res, 'You are not logined')) {
+            let user = await models.Users.findOne({
                 where: {
                     id: req.id
                 }
+            })
+            if (!user) {
+                res.status(404).json({
+                    msg: 'Account not define'
+                })
+            } else {
+                const old = user.img
+                if (user.img !== "public/images/default_avatar.jpg") {
+                    fs.unlink(old, (err) => {
+                        if (err) return console.log(err)
+                    });
+                }
+                await models.Users.update(
+                    {
+                        img: req.file.path
+                    },
+                    {
+                        where: {
+                            id: req.id
+                        }
+                    }
+                )
+                res.status(200).json({
+                    img: req.file.path
+                })
             }
-        )
-        res.status(200).json({
-            img: req.file.path
-        })
+        }
     } catch (error) {
         console.log(error)
     }
@@ -178,54 +162,49 @@ exports.image = async (req, res) => {
 }
 
 exports.edit = async (req, res) => {
-    if (!req.id) {
-        res.status(400).json({
-            msg: "You not authorizated"
-        })
-    }
-    let user = await models.Users.findOne({
-        where: {
-            id: req.id
-        }
-    })
-    if (!user) {
-        res.status(404).json({
-            msg: 'Account not define'
-        })
-    }
-    let body = {}
-    if (!!req.body.nick) {
-        const nickname = await models.Users.findOne({
-            where: {
-                nick: req.body.nick
-            }
-        })
-        if (!!nickname) {
-            res.status(400).json({
-                msg: "This nick already have"
+    if (check.variables(['id'], req, res, 'You are not logined')) {
+        if (check.variables(['nick', 'info', 'password'], req.body, res)) {
+            let user = await models.Users.findOne({
+                where: {
+                    id: req.id
+                }
             })
-        } else {
-            body["nick"] = req.body.nick
-        }
+            if (!user) {
+                res.status(404).json({
+                    msg: 'Account not define'
+                })
+            } else {
+                let body = {}
 
-    }
-    if (!req.body.info == false) {
-        body["info"] = req.body.info
-    }
-    if (!req.body.password == false) {
-        body["password"] = await bcrypt.hash(req.body.password, 10);
-    }
-    await models.Users.update(
-        body,
-        {
-            where: {
-                id: req.id
+                const nickname = await models.Users.findOne({
+                    where: {
+                        nick: req.body.nick
+                    }
+                })
+                if (!!nickname) {
+                    res.status(400).json({
+                        msg: "This nick already have"
+                    })
+                } else {
+                    body["nick"] = req.body.nick
+                    body["info"] = req.body.info
+                    body["password"] = await bcrypt.hash(req.body.password, 10);
+                    await models.Users.update(
+                        body,
+                        {
+                            where: {
+                                id: req.id
+                            }
+                        }
+                    )
+                    res.status(200).json({
+                        nick: req.body.nick,
+                        info: req.body.info,
+                    })
+                }
+
             }
         }
-    )
-    delete body['password']
-    res.status(200).json({
-        date: body
-    })
+    }
 }
